requirements.txt
attrs==19.3.0
enum34==1.1.10
hypothesis==4.57.1
PyHamcrest==1.10.1
six==1.14.0
sortedcontainers==2.1.0
